backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Data validation ────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Database ───────────────────────────────────────────────────
sqlalchemy==2.0.29
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# ── Queue & cache ──────────────────────────────────────────────
redis==5.0.4
celery==5.4.0
flower==2.0.1

# ── Object storage ─────────────────────────────────────────────
boto3==1.34.84

# ── Auth & security ────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.5

# ── HTTP client ────────────────────────────────────────────────
httpx==0.27.0
aiohttp==3.9.5

# ── File processing ────────────────────────────────────────────
python-magic==0.4.27
pillow==10.3.0
exifread==3.0.0

# ── Logging & observability ────────────────────────────────────
structlog==24.1.0
prometheus-client==0.20.0

# ── Utilities ──────────────────────────────────────────────────
python-dotenv==1.0.1
tenacity==8.3.0
orjson==3.10.3
